Now allow left and right modifier keys to be held down independently, this allows a maximum buy/sell of 5*5*20*20*100*100 = 2.5B! Also modified Format::Number to show "K" for numbers in the thousands, and also reduced the number of digits behind the decimal to 2 to allow large numbers to fit within the game display areas.

+++ source/HiringPanel.h
@@ -39,8 +39,8 @@ protected:
 private:
 	PlayerInfo &player;
 	
-	mutable int maxHire;
-	mutable int maxFire;
+	mutable int64_t maxHire;
+	mutable int64_t maxFire;
 };
